requirements.txt
arrow==0.12.1
backcall==0.1.0
beautifulsoup4==4.6.0
blessed==1.15.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
defusedxml==0.5.0
Django==2.0.6
django-admin-comments==1.0.1
django-auth-ldap==1.6.1
django-compat==1.0.15
django-crispy-forms==1.7.2
django-extensions==2.0.7
django-formtools==2.1
django-hijack==2.1.9
django-hijack-admin==2.1.9
django-model-utils==3.1.2
django-picklefield==1.0.0
django-python3-ldap==0.11.1
django-q==0.9.4
django-sslserver==0.20
future==0.16.0
humanize==0.5.1
idna==2.7
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.1
kerberos==1.3.0
ldap3==2.3
mysqlclient==1.3.13
oauthlib==2.1.0
parso==0.3.0
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.3.2
pyasn1-modules==0.2.1
pycryptodomex==3.6.3
Pygments==2.2.0
pyjwkest==1.4.0
PyJWT==1.6.4
python-dateutil==2.7.3
python-ldap==3.1.0
python3-openid==3.1.0
pytz==2018.4
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
simplegeneric==0.8.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
